refactor: 7. cleaning up initialization [TECH-344] (#709)
* refactor: AppLoader and init folder

* refactor: 8. Importing using created index files [TECH-344]  (#710)

* refactor: using index imports

* refactor: use index exports for loadingmasks
