+++ src/core_modules/capture-core/components/Buttons/ProgressButton.component.js
@@ -4,7 +4,7 @@ import { withStyles } from '@material-ui/core/styles';
 import green from '@material-ui/core/colors/green';
 
 import Button from './ButtonOld.component';
-import LoadingMaskForButton from '../LoadingMasks/LoadingMaskForButton.component';
+import { LoadingMaskForButton } from '../LoadingMasks';
 
 const styles = () => ({
     wrapper: {
