+++ src/core_modules/capture-core/components/LoadingMasks/index.js
@@ -1,3 +1,6 @@
 // @flow
 export { default as LoadingMask } from './LoadingMask.component';
 export { default as LoadingMaskForPage } from './LoadingMaskForPage.component';
+export { default as DialogLoadingMask } from './DialogLoadingMask.component';
+export { default as LoadingMaskElementCenter } from './LoadingMaskElementCenter.component';
+export { default as LoadingMaskForButton } from './LoadingMaskForButton.component';
