+++ src/index.js
@@ -1,6 +1,6 @@
 import 'capture-core-utils/extensions/asyncForEachArray';
 import 'capture-core-utils/extensions/arrayToHashMap';
 import './locales';
-import { initApp } from './init/entry';
+import { initApp } from './entry';
 
 initApp();
